+++ packages/cli/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Update inquirer dependencies (#2501)
+
+### Fixed
+- Not being able to select a custom repo when initializing a project (#2501)
 
 ## [5.1.0] - 2024-07-11
 ### Changed